fixed DAO useage inserting empty list of objects

+++ rest-server/engines/src/main/scala/com/actionml/engines/urnavhinting/URNavHintingDataset.scala
@@ -103,10 +103,12 @@ class URNavHintingDataset(engineId: String, val store: Store) extends Dataset[UR
           if(event.properties.getOrElse("conversion", false)) {
             // a conversion nav-event means that the active journey keyed to the user gets moved to the indicatorsDao
             val conversionJourney = activeJourneysDao.findMany(query = DaoQuery(filter = Seq(("entityId", event.entityId)))).toSeq
-            val taggedConvertedJourneys = conversionJourney.map(e => e.copy(conversionId = event.targetEntityId))
-            // tag these so they can be removed when the model is $deleted
-            indicatorsDao.insertMany(taggedConvertedJourneys)
-            activeJourneysDao.removeMany(("entityId", event.entityId))
+            if(conversionJourney.size != 0) {
+              val taggedConvertedJourneys = conversionJourney.map(e => e.copy(conversionId = event.targetEntityId))
+              // tag these so they can be removed when the model is $deleted
+              indicatorsDao.insertMany(taggedConvertedJourneys)
+              activeJourneysDao.removeMany(("entityId", event.entityId))
+            }
           } else {
             // save in journeys until a conversion happens
             activeJourneysDao.saveOne(event)
